[java1.8] Support building with java 8

+++ src/main/java/com/airbnb/airpal/core/health/PrestoHealthCheck.java
@@ -53,7 +53,7 @@ public class PrestoHealthCheck extends HealthCheck
                                     result = Iterables.getFirst(results, invalidValue);
                                     assert(result != null);
                                     assert(result.size() == 1);
-                                    assert(result.get(0) == 1);
+                                    assert((int)result.get(0) == 1);
                                 }
                                 client.advance();
                             }
